+++ gui/src/ui/update.rs
@@ -130,13 +130,13 @@ impl eframe::App for Gui {
 			diff
 		};
 		// If the queue is different from ours,
-		// recalculate the total `queue_runtime`.
+		// recalculate the total `queue_time`.
 		if self.kernel_returned && queue_diff {
-			self.queue_runtime = self.audio_state
+			self.queue_time = self.audio_state
 				.queue
 				.iter()
-				.map(|k| self.collection.songs[k].runtime.inner())
-				.sum::<u32>()
+				.map(|k| self.collection.songs[k].runtime.inner() as u64)
+				.sum::<u64>()
 				.into();
 		}